+++ src/main/gen/io/github/facilityapi/intellij/psi/FsdVisitor.java
@@ -1,0 +1,90 @@
+// This is a generated file. Not intended for manual editing.
+package io.github.facilityapi.intellij.psi;
+
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.PsiElement;
+
+public class FsdVisitor extends PsiElementVisitor {
+
+  public void visitAttribute(@NotNull FsdAttribute o) {
+    visitPsiElement(o);
+  }
+
+  public void visitAttributeList(@NotNull FsdAttributeList o) {
+    visitPsiElement(o);
+  }
+
+  public void visitAttributeParameter(@NotNull FsdAttributeParameter o) {
+    visitPsiElement(o);
+  }
+
+  public void visitDataSpec(@NotNull FsdDataSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitEnumSpec(@NotNull FsdEnumSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitEnumValue(@NotNull FsdEnumValue o) {
+    visitPsiElement(o);
+  }
+
+  public void visitEnumValueList(@NotNull FsdEnumValueList o) {
+    visitPsiElement(o);
+  }
+
+  public void visitErrorList(@NotNull FsdErrorList o) {
+    visitPsiElement(o);
+  }
+
+  public void visitErrorSetSpec(@NotNull FsdErrorSetSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitErrorSpec(@NotNull FsdErrorSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitField(@NotNull FsdField o) {
+    visitPsiElement(o);
+  }
+
+  public void visitMarkdownRemarks(@NotNull FsdMarkdownRemarks o) {
+    visitPsiElement(o);
+  }
+
+  public void visitMethodSpec(@NotNull FsdMethodSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitReferenceType(@NotNull FsdReferenceType o) {
+    visitPsiElement(o);
+  }
+
+  public void visitServiceItems(@NotNull FsdServiceItems o) {
+    visitPsiElement(o);
+  }
+
+  public void visitServiceSpec(@NotNull FsdServiceSpec o) {
+    visitPsiElement(o);
+  }
+
+  public void visitType(@NotNull FsdType o) {
+    visitPsiElement(o);
+  }
+
+  public void visitTypeIdentifier(@NotNull FsdTypeIdentifier o) {
+    visitNamedElement(o);
+  }
+
+  public void visitNamedElement(@NotNull FsdNamedElement o) {
+    visitPsiElement(o);
+  }
+
+  public void visitPsiElement(@NotNull PsiElement o) {
+    visitElement(o);
+  }
+
+}